+++ src/index.ts
@@ -1,8 +1,8 @@
-import type { RegexElement } from './components/types';
-import { encodeSequence } from './encoder/encoder';
-
 export type * from './components/types';
 
+export { buildPattern, buildRegex } from './builders';
+
+export { capture } from './components/capture';
 export {
   any,
   anyOf,
@@ -18,25 +18,3 @@ export {
   zeroOrMore,
 } from './components/quantifiers';
 export { repeat } from './components/repeat';
-
-/**
- * Generate RegExp object for elements.
- *
- * @param elements
- * @returns
- */
-export function buildRegex(...elements: Array<RegexElement | string>): RegExp {
-  const pattern = encodeSequence(elements).pattern;
-  return new RegExp(pattern);
-}
-
-/**
- * Generate regex pattern for elements.
- * @param elements
- * @returns
- */
-export function buildPattern(
-  ...elements: Array<RegexElement | string>
-): string {
-  return encodeSequence(elements).pattern;
-}
